Do not deep-copy tables with metatables.
A table with a metatable is almost always going to represent some kind
of instance, which should not be copied. A table without a metatable is
almost always structural, and therefore should be copied. If the user
needs to force a behavior, they can just use a function. The result of a
function is not interpreted any further, so it can be used for absolute
control over the initial value.