Fix naming of signals and events.
An "event" is an element of a class that is fired when something
happens. A "signal" is an actual value that encapsulates the connection
behavior of an event.

This change swaps the names of Signal and Event to match these
definitions. Signal is equivalent to RBXScriptSignal, and Event is
equivalent to BindableEvent.

The flaw in The Roblox API is with BindableEvent.Event, which should
really be named BindableEvent.Signal.