Attempt to copy breakpoints if debugger is available.
When a module is copied, any breakpoints attached to it are not copied.
If the current environment can access the DebuggerManager, then it
attempts to copy any breakpoints attached to the original and assigns
them to the copy.

In practice, DebuggerManager seems to work poorly due to recent changes
to the debugger. Maybe revert.