Refactor Binstruct.
Implements dynamically sized arrays. Size is determined by the specified
field of the parent struct.

A codec now contains a single list of instructions. Each instruction has
a number of "columns", which are used to implement differences in
decoding and encoding. The code for decoding and encoding are now the
same, the only difference being which instruction column is selected.

Instructions are implemented as separate functions selected by op.

The processor has a distinct set of registers contained within a table.
A subset of these registers are copied when a stack frame is pushed or
popped.

Adds FORC, FORF, and JMPN instructions, used to implement the array
type.

Adds a Dump method to display the instructions generated for the codec.