Fix scoping to separate transformed and structural values.
The decoding filter is delayed to the POP instruction, so that scoping
operates on the structural value, rather than the transformed value.
Assignment of the value to its parent field is also moved out of PUSH
and into POP.

Argument 1 of PUSH must be a function that prepares a structural value
to scope into. Argument 1 of POP must be a function that transforms the
structural value after it has been scoped out, then assigns the
result.

Argument 2 of PUSH must be a function that transforms the value back
into a structural value. Argument 2 of POP is unused, because no values
need to be assigned.

This fix enables more complex filtering. For example, with the right
filters, encoding and decoding Vector3 types directly becomes possible.