Refactor jumping.
FORx instructions failed to check their condition at the start of the
loop, resulting in arrays always containing at least one element.

This change refactors FORx to jump to the end of the loop if needed.
Additionally, the JA register is removed, and JMPN jumps directly to a
given address.

Note: it is no longer safe to arbitrarily insert/remove instructions;
only appending is safe unless special care is taken.