Change instance finalizer to set Parent to nil.
ClearAllChildren just calls Destroy on each child (thanks Roblox). It
could have been changed to Remove, which has the desired behavior, but
I decided that dissolving the tree is not necessarily desirable.
Instead, the finalizer will do the bare minimum. If the user wants to
call Remove, then they can do it through a function task.