Implement calculated vector size.
If vector size is a function, that function is used to calculate the
size. The function is similar to a hook, receiving a stack function
and the global table.

For example, this can be used to determine the size of a flat
multidimensional array.

Adds the FORX instruction, which sets N to the result of the parameter
function.