Improve union constructor.
Arguments form pairs. The first element of a pair is a Hook, the second
element is a TypeDef. A copy of the TypeDef is made, the hook is set to
it, and then it is added to the union. If a TypeDef is nil, then 0-size
padding is used instead.