Move global generation to parseDef.
Similar to hooks, generation of a GLOBAL instruction is always done last
for each type. Therefore it can be moved to parseDef.

pad and align types can now have a global defined, but don't set a
value. To prevent the current TABLE[KEY] from being set to the global
incorrectly, pad and align set KEY to nil.