Rewrite union as conditional statement.
Instead of evaluating hooks on each subtype, each entry to a union is a
clause, similar to a struct field. Each clause has a Calc expression.
If an expression returns true, then the associated definition is used,
and no more expressions are evaluated.

If an expression is the boolean true instead of a function, then it
always evaluates to true. This can be used to implement an "else"
clause more easily.

- Adds EXPR instruction, which has both an address and an exit address.
  The exit address, which indicates where the end of the conditional
  statement is located. This is used to terminate the statement when a
  previous clause evaluated to true.
- Adds UXPR instruction, which does the same thing, except that it
  unconditionally evaluates to true.