Add ptr type.
A ptr is nothing more than a table that contains another type. The
compiler resolves it simply by compiling the referred type. Causes an
error if the referent is not a TypeDef (null pointers are not
allowed).

This allows definitions to use a type before it is defined.

Note that ptr is not a regular TypeDef. It has no extra hook or global
fields. Rather, parseDef will parse a ptr immediately.

To prevent stack overflows, a ptr cannot be dereferenced while it is
already being dereferenced.