Distinguish between serial and concurrent units.
By default, units now run serially. This fixes an issue where concurrent
units share upvalues, producing race conditions.

Units have an option to run concurrently. In a later step, the user will
be able to mark a unit as concurrent. In this case, it will be assumed
by the framework that relevant upvalues are accessed safely.